Clarify semantics of publish confirms
Looking at the code it looks more that I should look at r.isAck() to tell if the message is confirmed by the broker.
So is it true that the Flux<OutboundMessageResult> contains an entry for every outbound message regardless if confirmed or not.
Or does it only contain an element for every successful delivery?