Add failure-proof channel proxy for resource creation
The proxy creates a new channel delegate if the current one is closed,
no matter the reason. This makes Sender a bit more robust.

A channel is typically closed because of an error. Here the error will
bubble up to let the user know about it. But the Sender will still be
usable after.

References #21