Using BaseSubscriber in Mono.subscribe() method of each consume method so the exception thrown in consumer is then handled by errorConsumer.
The problem was discovered in our project when Receiver was consuming a queue that didn't exist. This is the same issue as described in https://github.com/reactor/reactor-core/issues/1995 and usage of BaseSubscriber is a workaround.

(cherry picked from commit 28db7d5107ba7f03c4c735a9f5df82c1bb99adfc)