[chores] Copyright: Correct start year for all files (#162)
This is a follow up on #161 which mistakenly harmonized all the start
dates in range to 2017. Some files have more ancient history than that,
having been transferred from older repositories. Others are more recent.

This commit manually reverts the start year part of the previous change
to refine the start dates as much as possible.

See #161.
See reactor/reactor#682.